pp: Move one_shot to pandaproxy
This will be used by pandaproxy/rest to defer startup until first
request.
